+++ src/app/components/Filter.tsx
@@ -3,11 +3,11 @@ import FilterAltIcon from "@mui/icons-material/FilterAlt";
 import IconButton from "@mui/material/IconButton";
 
 function Filter() {
-  return (
-    <IconButton size="large" color="inherit">
-      <FilterAltIcon />
-    </IconButton>
-  );
+    return (
+        <IconButton size="large" color="inherit">
+            <FilterAltIcon />
+        </IconButton>
+    );
 }
 
 export default Filter;
