+++ src/app/components/MovieCard.tsx
@@ -11,74 +11,72 @@ import { useQuery } from "react-query";
 import { useNavigate } from "react-router-dom";
 import DetailsPage from "../pages/DetailsPage";
 
-function MovieCard() {
-  let navigate = useNavigate();
+function MovieCard(props: any) {
+    const {filter} = props;
 
-  function navigateToDetails(movieId: any) {
-    navigate("/DetailsPage", { state: { movieId } });
-  }
+    let navigate = useNavigate();
 
-  const apiUrl =
-    "https://wi2020seb-cinema-api-dev.azurewebsites.net/movie/getAll";
+    function navigateToDetails(movieId: any) {
+        navigate("/DetailsPage", { state: { movieId } });
+    }
 
-  const { isLoading, error, data } = useQuery("Movies", () =>
-    fetch(apiUrl).then((res) => res.json())
-  );
+    const apiUrl =
+        "https://wi2020seb-cinema-api-dev.azurewebsites.net/movie/getAll";
 
-  const [filter, setFilter] = useState("");
+    const { isLoading, error, data } = useQuery("Movies", () =>
+        fetch(apiUrl).then((res) => res.json())
+    );
 
-  const handleSearchChange = (e: any) => {
-    setFilter(e.target.value);
-  };
 
-  return (
-    <Container
-      sx={{
-        bgcolor: "background.paper",
-        pt: 8,
-        pb: 6,
-        position: "relative",
-      }}
-      maxWidth="md"
-    >
-      <Grid container spacing={4}>
-        {data?.map((movie: any) => (
-          <Grid item key={movie.id} xs={12} sm={6} md={4}>
-            <Card
-              sx={{ height: "100%", display: "flex", flexDirection: "column" }}
-            >
-              <CardMedia
-                component="img"
-                sx={{
-                  pt: "6%",
-                }}
-                image={movie.pictureLink}
-                alt="poster"
-              />
-              <CardContent sx={{ flexGrow: 1 }}>
-                <Typography gutterBottom variant="h5" component="h2">
-                  {movie.titel}
-                </Typography>
-                <Typography>
-                  {movie.duration} Min FSK {movie.fsk}
-                </Typography>
-              </CardContent>
-              <CardActions>
-                <Button
-                  size="small"
-                  onClick={() => {
-                    navigateToDetails(`${movie.id}`);
-                  }}
-                >
-                  Tickets
-                </Button>
-              </CardActions>
-            </Card>
-          </Grid>
-        ))}
-      </Grid>
-    </Container>
-  );
+    return (
+        <Container
+            sx={{
+                bgcolor: "background.paper",
+                pt: 8,
+                pb: 6,
+                position: "relative",
+            }}
+            maxWidth="md"
+        >
+            <Grid container spacing={4}>
+                {data?.map((movie: any) => (
+                    movie.titel.toLowerCase().includes(filter) &&
+                    <Grid item key={movie.id} xs={12} sm={6} md={4}>
+                        <Card
+                            sx={{ height: "100%", display: "flex", flexDirection: "column" }}
+                        >
+                            <CardMedia
+                                component="img"
+                                sx={{
+                                    pt: "6%",
+                                }}
+                                image={movie.pictureLink}
+                                alt="poster"
+                            />
+                            <CardContent sx={{ flexGrow: 1 }}>
+                                <Typography gutterBottom variant="h5" component="h2">
+                                    {movie.titel}
+                                </Typography>
+                                <Typography>
+                                    {movie.duration} Min FSK {movie.fsk}
+                                </Typography>
+                            </CardContent>
+                            <CardActions>
+                                <Button
+                                    size="small"
+                                    onClick={() => {
+                                        navigateToDetails(`${movie.id}`);
+                                    }}
+                                >
+                                    Tickets
+                                </Button>
+                            </CardActions>
+                        </Card>
+                    </Grid>
+                ))}
+            </Grid>
+        </Container>
+    );
 }
 
 export default MovieCard;
