--- src/app/components/filter.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import FilterAltIcon from "@mui/icons-material/FilterAlt";
-import IconButton from "@mui/material/IconButton";
-
-function Filter() {
-  return (
-    <IconButton size="large" color="inherit">
-      <FilterAltIcon />
-    </IconButton>
-  );
-}
-
-export default Filter;
